requirements.txt
alembic==1.0.11
appnope==0.1.0
astroid==2.2.5
backcall==0.1.0
bcrypt==3.1.7
CacheControl==0.12.5
cachetools==3.1.1
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
Click==7.0
decorator==4.4.0
feedparser==5.2.1
firebase-admin==2.17.0
Flask==1.1.1
Flask-Cors==3.0.8
Flask-Login==0.4.1
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.2
flatten-json==0.1.7
gcloud==0.17.0
google-api-core==1.14.2
google-api-python-client==1.7.10
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-cloud-core==1.0.3
google-cloud-firestore==1.4.0
google-cloud-storage==1.18.0
google-resumable-media==0.3.2
googleapis-common-protos==1.6.0
grpcio==1.22.0
gunicorn==19.9.0
httplib2==0.13.1
idna==2.8
ipython==7.7.0
ipython-genutils==0.2.0
isort==4.3.21
itsdangerous==1.1.0
jedi==0.15.1
Jinja2==2.10.1
jws==0.1.3
lazy-object-proxy==1.4.1
Mako==1.1.0
MarkupSafe==1.1.1
mccabe==0.6.1
msgpack==0.6.1
oauth2client==3.0.0
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
prompt-toolkit==2.0.9
protobuf==3.9.1
ptyprocess==0.6.0
pyasn1==0.4.6
pyasn1-modules==0.2.6
pycparser==2.19
pycryptodome==3.4.3
Pygments==2.4.2
pylint==2.3.1
Pyrebase==3.0.27
python-dateutil==2.8.0
python-editor==1.0.4
python-jwt==2.0.1
pytz==2019.2
redis==3.3.8
requests==2.22.0
requests-toolbelt==0.7.0
rsa==4.0
schedule==0.6.0
six==1.12.0
SQLAlchemy==1.3.6
traitlets==4.3.2
uritemplate==3.0.0
urllib3==1.25.3
wcwidth==0.1.7
Werkzeug==0.15.5
wrapt==1.11.2
WTForms==2.2.1
